requirements.txt
google-genai
numpy
tiktoken
